feat: implements the server action to form-login.
