+++ apps/web/components/login-page.tsx
@@ -1,0 +1,29 @@
+import { cn } from "@/lib/utils";
+import {
+  Card,
+  CardContent,
+  CardDescription,
+  CardHeader,
+  CardTitle,
+} from "@/components/ui/card";
+import FormLogin from "./ui/form-login";
+
+export function LoginPage({
+  className,
+  ...props
+}: React.ComponentProps<"div">) {
+  return (
+    <div className={cn("flex flex-col gap-6", className)} {...props}>
+      <Card>
+        <CardHeader>
+          <CardTitle>Login com sua conta</CardTitle>
+          <FormLogin />
+          <CardDescription>
+            Digite seu e-mail abaixo para acessar sua conta
+          </CardDescription>
+        </CardHeader>
+        <CardContent></CardContent>
+      </Card>
+    </div>
+  );
+}
